requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
python-dotenv==1.2.1
psycopg2-binary
gunicorn

